+++ model/model_internal_test.go
@@ -1,0 +1,7 @@
+package model
+
+func (m *Measurement) MakeGenericTestKeysEx(
+	marshal func(v interface{}) ([]byte, error),
+) (map[string]interface{}, error) {
+	return m.makeGenericTestKeys(marshal)
+}
