+++ model/model_test.go
@@ -139,3 +139,42 @@ func TestPrivacySettingsApplyMarshalError(t *testing.T) {
 		t.Fatal("expected an error here")
 	}
 }
+
+func TestMakeGenericTestKeysIdempotent(t *testing.T) {
+	m := new(model.Measurement)
+	m.TestKeys = make(map[string]interface{})
+	_, err := m.MakeGenericTestKeysEx(
+		func(interface{}) ([]byte, error) {
+			return nil, errors.New("mocked error")
+		},
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+func TestMakeGenericTestKeysSuccess(t *testing.T) {
+	m := makeMeasurement("127.0.0.1", "AS137", "IT")
+	out, err := m.MakeGenericTestKeys()
+	if err != nil {
+		t.Fatal(err)
+	}
+	if out["client_resolver"].(string) != "91.80.37.104" {
+		t.Fatal("expected different client resolver here")
+	}
+}
+
+func TestMakeGenericTestKeysMarshalError(t *testing.T) {
+	m := new(model.Measurement)
+	out, err := m.MakeGenericTestKeysEx(
+		func(interface{}) ([]byte, error) {
+			return nil, errors.New("mocked error")
+		},
+	)
+	if err == nil {
+		t.Fatal("expected an error here")
+	}
+	if out != nil {
+		t.Fatal("expected nil output here")
+	}
+}
