+++ testdata/.gitignore
@@ -7,3 +7,4 @@
 /oonipsiphon
 /psiphon_config.json
 /psiphon_unit_tests/
+/test-download-resources-idempotent*/
